+++ thewalrus/tests/test_hermite_multidimensional.py
@@ -225,7 +225,9 @@ def test_multi_cutoffs_multidim_herm_numba():
     y = np.random.rand(d) + 1j * np.random.rand(d)
     C = 0.5
 
-    n = [np.random.randint(0, 4) for _ in range(d)]
-    poly = hermite_multidimensional_numba(R, [3+n[0], 3+n[1], 3+n[2], 3+n[3]], y, C)
+    cutoffs = [np.random.randint(3, 7) for _ in range(d)]
+    poly = hermite_multidimensional_numba(R, cutoffs, y, C)
     poly_expected = hermite_multidimensional_numba(R, 3, y, C)
+    assert np.allclose(poly[:3, :3, :3, :3], poly_expected)
+    poly_expected = hermite_multidimensional_numba(R, np.array(3), y, C) # testing ndarrary cutoff int
     assert np.allclose(poly[:3, :3, :3, :3], poly_expected)