+++ .github/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Improvements
 
 ### Bug fixes
-
+* `hermite_multidimensional_numba` now can handle a cutoff of type `np.ndarray` with `shape=[]`. [#283](https://github.com/XanaduAI/thewalrus/pull/283)
 ### Breaking changes
 
 ### Contributors